Revert "The constructor CmdLineException(String) is deprecated"
This reverts commit 6b05a4f7b3386b7b946b8301736d2fa32feea3cd [formerly 07f99362575f324e466d7f5ae49d5f8f69cf7a8a].

07f9 seems to require a more recent args4j, but I cannot locate a CQ
that verifies we can use a version more recent then 2.0.12. 2.0.16
has been released, but the Hudson CI instance at Eclipse won't
build it.

Since the commit fixes an issue identified in March but wasn't
actually submitted to the tree until September, we can continue to
ignore whatever the problem is/was until someone can attempt a more
correctly working solution.

Change-Id: I94fa432c219bda21c1126976bb60e5292760092e

Former-commit-id: 31261b5c2f261653e11d0d9960ce26fa0d4c82d5