Revert "Stop PathFilter after walking all matching paths"
This reverts commit e99c348e1a75bfc4ca79596669066598c28c1977 [formerly 75eb6a147ffa8089e48d409f0215a16c755305e7].
Applications that want a PathFilter to abort the walk early should be
using PathFilterGroup. When a PathFilterGroup is created with exactly
one path its implementation is the same that 75eb6 tried to perform,
but has been long documented as having the behavior of breaking a
higher level OR filter graph node.

Change-Id: I6c85d75c474784471c32e866eef3402b9f193c08

Former-commit-id: 0b6387fe7c575b40578a8478392eb1cf35f8a136